update Reservation details
