+++ src/pages/UserProfilePage.js
@@ -6,7 +6,7 @@ import PersonalDetails from "../components/PersonalDetails";
 import Header from "../components/Header";
 import Footer from "../components/Footer";
 import UserProfileNav from "../components/UserProfileNav";
-import ReservationDetails from "../pages/ReservationDetails";
+import ReservationDetails from "../components/ReservationDetails";
 
 export default function UserProfilePage() {
   const theme = useTheme();
@@ -25,7 +25,7 @@ export default function UserProfilePage() {
               <Route path={`/personal`} element={<PersonalDetails />} />
               <Route path={`/reservations`} element={<Reservations />} />
               <Route
-                path={`/reservations/:id`}
+                path={`/reservations/:reservationId`}
                 element={<ReservationDetails />}
               />
             </Routes>
